requirements.txt
alembic==1.14.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docutils==0.21.2
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Kivy==2.3.0
Kivy-Garden==0.1.5
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
psycopg2-binary==2.9.10
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
paho-mqtt==2.1.0